Up number of connections to mysqlchk
We were experiencing haproxies marking DB hosts down due to checks
failing. The root cause was a limit on number of connections xinetd
server allows; when this is exceeded xinetd returns a RST which leads
haproxy to thinking the DB is down.

The default is 50 connections in 10 secs which is too low. Up it to 100
connections in 1 sec.

Change-Id: I05ed2f41701bd5c709566cf026d266c29348d7cf