Fix wrong version of mariadb being installed
Under certain circumstances, the wrong version of mariadb can be
installed. One such case is when the initial apt::key cannot be
installed.

 Error: Could not set 'present' on ensure: could not resolve
 https://supplychain.mariadb.com/MariaDB-Server-GPG-KEY (file:
 /etc/puppetlabs/code/modules/apt/manifests/key.pp, line: 54)

This error does not prevent mariadb repo, or packages from being
installed. However, the packages will not be from the repo in this
class, as the repo is not valid without a key.

Fix this by making apt::source require apt::key. Also make a few minor
changes:

1. Change the apt::pin dependency to a require. This is recommended by
   Puppet [1]
2. Use a notify to Exec['apt_update'] instead of Class['apt::update'].
   This is following puppetlabs-apt docs [2]

[1]: https://www.puppet.com/docs/puppet/7/lang_relationships.html#lang_rel_chaining_arrows
[2]: https://forge.puppet.com/modules/puppetlabs/apt/8.4.0/readme#update-the-list-of-packages

Change-Id: If3c213e0067e1dd3c195349791ce4fcc60cb7419